backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
alembic>=1.14
asyncpg>=0.30
pydantic>=2.10
pydantic-settings>=2.7
python-jose[cryptography]>=3.4
passlib[bcrypt]>=1.7
httpx>=0.28
pytest>=8.3
pytest-asyncio>=0.25
ruff>=0.9
mypy>=1.14
types-python-jose>=3.3.4
openai>=1.65
